--- app/src/main/java/com/example/corner_library/model/Category.kt
@@ -1,10 +1,0 @@
-package com.example.corner_library.model
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class Category(
-    val title: String,
-    val projectList: List<Project>
-) : Parcelable
