--- app/src/main/java/com/example/corner_library/model/Tag.kt
@@ -1,12 +1,0 @@
-package com.example.corner_library.model
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class Tag(
-    val id: Number,
-    val name: String,
-) : Parcelable {
-
-}
